No guess algorithm works for all but number 4
Needs troubleshooting for number 4. 3 doesn't work as the correct center isn't found as a possible center, but it is close. 
The code needs to be further condensed for ease of use. We should have the user only input one file name. We should make a method that takes the file path name and automatically puts it in the numpy formatting.
Also, find a way to make many variables in a loop so we don't have to make 16 points.
We may reduce the image quality to improve run time.