e to pi
The out-of-image bound values were called e and the literal values in the image that were too low were called e as well. The latter of the two is now called pi and pi is removed from the "point amount" calculation. But now these values are not ending the point expanding outwards.